refactor: Singular naming for tenant context.
