+++ src/App.tsx
@@ -6,7 +6,7 @@ import { theme, Skeleton } from '@tabetalt/kit';
 import { getRoutes } from './routing';
 import gqlClient from './api/client';
 import { useAuth } from './context/AuthContext';
-import { TenantsProvider } from './context/TenantsContext';
+import { TenantProvider } from './context/TenantContext';
 
 const renderLoader = () => <Skeleton />;
 
@@ -17,9 +17,9 @@ export const App: React.FC = () => {
   return (
     <ApolloProvider client={gqlClient}>
       <ThemeProvider theme={theme}>
-        <TenantsProvider>
+        <TenantProvider>
           <Suspense fallback={renderLoader()}>{routes}</Suspense>
-        </TenantsProvider>
+        </TenantProvider>
       </ThemeProvider>
     </ApolloProvider>
   );
